requirements.txt
# Core web app
streamlit>=1.35.0

# Language Models / RAG
langchain>=0.2.0
pinecone-client>=3.0.0

# Audio: speech-to-text & text-to-speech
gTTS>=2.5.0
SpeechRecognition>=3.10.0
transformers>=4.41.0
torch>=2.0.0
torchaudio>=2.0.0

# Visual Question Answering (BLIP)
sentencepiece>=0.1.99
accelerate>=0.28.0

# PDF processing
PyPDF2>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# (Optional) if using Llama or Ollama models locally
llama-cpp-python>=0.2.0
